refactor: extract FiberRootNode to a new file
